Test coverage and commits needed to support that
